requirements.txt
requests
python-dotenv
pandas
numpy
mlxtend
sqlite3